chore(review): address Copilot comments on PR #306
Three small cosmetic / accuracy fixes from Copilot review:

1. AppointmentController: drop the redundant `(int)` cast on
   `Math.clamp(size, 1, 200)`. With int args the (long, int, int)
   overload is selected, which already returns int.

2. AppointmentServiceImpl: the helper-extraction comment claimed the
   resulting methods are "unit-testable", but they are `private static`
   — not directly testable from a test class. Drop the false claim;
   keep the accurate "independently readable" part.

3. AppointmentServiceImpl: two helper Javadocs claimed the extraction
   is "byte-for-byte equivalent" to the original nested ternaries.
   Extracting into separate methods changes the generated bytecode
   (extra invoke, different stack layout) — soften to "behaviour
   equivalent" so the doc isn't overpromising at the JVM level.

No behaviour change.

+++ hospital-core/src/main/java/com/example/hms/controller/AppointmentController.java
@@ -182,7 +182,7 @@ public class AppointmentController {
         // Math.min(Math.max(size, 1), 200) idiom for clarity.
         // Math.max(page, 0) is still needed because page has no upper
         // bound — only a floor.
-        Pageable pageable = PageRequest.of(Math.max(page, 0), (int) Math.clamp(size, 1, 200), Sort.by(Sort.Direction.DESC, SORT_APPOINTMENT_DATE));
+        Pageable pageable = PageRequest.of(Math.max(page, 0), Math.clamp(size, 1, 200), Sort.by(Sort.Direction.DESC, SORT_APPOINTMENT_DATE));
         Page<AppointmentResponseDTO> resultPage = appointmentService.searchAppointments(
             filter, pageable, locale, getUsername(authentication));
         HttpHeaders headers = new HttpHeaders();
